% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcaf.r
\docType{data}
\name{root_path}
\alias{root_path}
\title{data loader}
\format{An object of class \code{character} of length 1.}
\usage{
root_path
}
\description{
This package allows you to load Cryptocurrency data
}
\keyword{datasets}
